several code quality improvements
use int instead of Long for correction round
use helper methods to avoid code duplication
clean the API of Submission.java

+++ src/main/java/de/tum/in/www1/artemis/repository/ResultRepository.java
@@ -111,7 +111,7 @@ public interface ResultRepository extends JpaRepository<Result, Long> {
                                 AND NOT EXISTS (select prs from p.results prs where prs.assessor.id = p.student.id)
                             ) >= (:correctionRound + 1L)
             """)
-    long countNumberOfFinishedAssessmentsByCorrectionRoundsAndExerciseIdIgnoreTestRuns(@Param("exerciseId") Long exerciseId, @Param("correctionRound") Long correctionRound);
+    long countNumberOfFinishedAssessmentsByCorrectionRoundsAndExerciseIdIgnoreTestRuns(@Param("exerciseId") Long exerciseId, @Param("correctionRound") int correctionRound);
 
     @EntityGraph(type = LOAD, attributePaths = { "feedbacks" })
     List<Result> findAllWithEagerFeedbackByAssessorIsNotNullAndParticipation_ExerciseIdAndCompletionDateIsNotNull(Long exerciseId);
