+++ src/main/java/de/tum/in/www1/artemis/service/ParticipationService.java
@@ -1007,7 +1007,7 @@ public class ParticipationService {
      * @param correctionRound - the correction round we want our submission to have results for
      * @return a list of participations including their submitted submissions that do not have a manual result
      */
-    public List<StudentParticipation> findByExerciseIdWithLatestSubmissionWithoutManualResultsAndNoTestRun(Long exerciseId, Long correctionRound) {
+    public List<StudentParticipation> findByExerciseIdWithLatestSubmissionWithoutManualResultsAndNoTestRun(Long exerciseId, int correctionRound) {
         return studentParticipationRepository.findByExerciseIdWithLatestSubmissionWithoutManualResultsAndNoTestRunParticipation(exerciseId, correctionRound);
     }
